--- src/factories/asset.factory.ts
@@ -1,72 +1,0 @@
-import { Asset, AssetDynamic } from "..";
-
-const defaultAssetStaticVaultV2: Asset<"VAULT_V2"> = {
-  address: "0x001",
-  typeId: "VAULT_V2",
-  token: "0x001",
-  name: "ASSET",
-  version: "1",
-  symbol: "ASS",
-  decimals: "18",
-  tokenId: "0x001",
-  underlyingTokenBalance: {
-    amount: "1",
-    amountUsdc: "1"
-  },
-  metadata: {
-    symbol: "str",
-    pricePerShare: "Int",
-    migrationAvailable: true,
-    latestVaultAddress: "0x001",
-    depositLimit: "Int",
-    emergencyShutdown: true,
-
-    controller: "0x001",
-    totalAssets: "Int",
-    totalSupply: "Int",
-
-    displayName: "str",
-    displayIcon: "str",
-    defaultDisplayToken: "0x001",
-
-    hideIfNoDeposits: true
-  }
-};
-
-const defaultAssetDynamicVaultV2: AssetDynamic<"VAULT_V2"> = {
-  address: "0x001",
-  typeId: "VAULT_V2",
-  tokenId: "0x001",
-  underlyingTokenBalance: {
-    amount: "1",
-    amountUsdc: "1"
-  },
-  metadata: {
-    symbol: "str",
-    pricePerShare: "Int",
-    migrationAvailable: true,
-    latestVaultAddress: "0x001",
-    depositLimit: "Int",
-    emergencyShutdown: true,
-
-    controller: "0x001",
-    totalAssets: "Int",
-    totalSupply: "Int",
-
-    displayName: "str",
-    displayIcon: "str",
-    defaultDisplayToken: "0x001",
-
-    hideIfNoDeposits: true
-  }
-};
-
-export const createMockAssetStaticVaultV2 = (overwrites: Partial<Asset<"VAULT_V2">> = {}) => ({
-  ...defaultAssetStaticVaultV2,
-  ...overwrites
-});
-
-export const createMockAssetDynamicVaultV2 = (overwrites: Partial<AssetDynamic<"VAULT_V2">> = {}) => ({
-  ...defaultAssetDynamicVaultV2,
-  ...overwrites
-});
