--- src/factories/token.factory.ts
@@ -1,17 +1,0 @@
-import { Token } from "..";
-
-export const defaultToken: Token = {
-  address: "0x001",
-  decimals: "18",
-  symbol: "DEAD",
-  name: "Dead Token",
-  priceUsdc: "0",
-  supported: {
-    zapper: true
-  }
-};
-
-export const createMockToken = (overwrites: Partial<Token> = {}) => ({
-  ...defaultToken,
-  ...overwrites
-});
