+++ src/test-utils/factories/index.ts
@@ -1,3 +1,6 @@
 export * from "./asset.factory";
 export * from "./balance.factory";
+export * from "./earningsAssetData.factory";
+export * from "./earningsUserData.factory";
 export * from "./token.factory";
+export * from "./tokenBalance.factory";
